ワザ相性による無効化のチェック位置を修正 ポケモン捕獲後のジングル再生終了後、通常の戦闘勝利ＢＧＭが流れるようにした。 バトルから天候「きり」を削除した。

+++ prog/src/battle/btl_util.h
@@ -150,7 +150,7 @@ static inline BPP_SICK_CONT BPP_SICKCONT_MakeTurn( u8 turns )
 /*
  *  ターン数型（パラメータ付き）作成
  */
-static inline BPP_SICK_CONT BPP_SICKCONT_MakeTurnParam( u8 turns, u8 param )
+static inline BPP_SICK_CONT BPP_SICKCONT_MakeTurnParam( u8 turns, u16 param )
 {
   BPP_SICK_CONT  cont;
   cont.raw = 0;
